Fix: set policy on trust
Motivation:
The configured policy for each trust object was not used.
This means that each revocation check in
OCSPAuthURLSessionDelegate would use whatever policy was
previously set on the trust object -- likely the system
defaults. OCSPCache should not be used before this commit.

Modifications:
- Set policy in `evaluateWithPolicy`

Result:
Revocation checks are performed as expected.