Fix: retain all policies set on trust
Motivation:
This fixes a bug which resulted from the use of
SecTrustSetPolicies. This call overwrites the policies
set on the trust. This resulted in the default policy
for checking the hostname of a certificate being removed
in the case of NSURLSession.

Modifications:
- Append the OCSP policy to the existing policies

Result:
Fixes a major bug where hostname validation was erroneously
removed at runtime.